SIB-141 clean up acumulated error advice into one file
