+++ src/main/java/com/capgemini/fs/coindashboard/controller/exceptionHandler/ErrorHandlingAdvice.java
@@ -7,7 +7,14 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ControllerAdvice
-public class InvalidPathAdvice {
+public class ErrorHandlingAdvice {
+
+  @ResponseBody
+  @ExceptionHandler(CoinNotFoundException.class)
+  @ResponseStatus(HttpStatus.NOT_FOUND)
+  String coinNotFoundHandler(CoinNotFoundException ex) {
+    return ex.getMessage();
+  }
 
   @ResponseBody
   @ExceptionHandler(InvalidPathException.class)
@@ -15,4 +22,13 @@ public class InvalidPathAdvice {
   String coinNotFoundHandler(InvalidPathException ex) {
     return ex.getMessage();
   }
+
+  @ResponseBody
+  @ExceptionHandler(ServerIsNotRespondingException.class)
+  @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+  String ServerNotRespondingHandler(ServerIsNotRespondingException ex) {
+    return ex.getMessage();
+  }
+
+
 }
